Move git::HostingProvider in separate module
